Created return to top button
Created a button that when pressed will return the user to the top of the window.

The button will only appear when the user has scrolled down more than a defined amount of pixels (currently 100) and it will be fixed at the top middle of the window.

This resolves issue #64